@@ -1,6 +1,6 @@
 {
   "name": "react-redux-firebase",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "description": "Redux integration for Firebase. Comes with a Higher Order Component for use with React.",
   "main": "dist/index.js",
   "module": "src/index.js",
@@ -14,7 +14,7 @@
     "lint": "eslint src/**",
     "lint:fix": "npm run lint -- --fix",
     "test": "mocha -R spec --compilers js:babel-core/register ./test/setup.js ./test/**/*.spec.js",
-    "test:cov": "babel-node $(npm bin)/isparta cover --report lcov ./node_modules/mocha/bin/_mocha -- ./test/** --recursive",
+    "test:cov": "istanbul cover ./node_modules/mocha/bin/_mocha -- ./test/** --recursive --report lcov --compilers js:babel-register --require babel-polyfill",
     "codecov": "cat coverage/*/lcov.info | codecov",
     "prepush": "npm run lint:fix"
   },
@@ -39,10 +39,10 @@
     }
   ],
   "dependencies": {
-    "firebase": "^3.5.3",
+    "firebase": "^3.6.1",
     "immutable": "^3.8.1",
     "jwt-decode": "^2.1.0",
-    "lodash": "^4.16.6"
+    "lodash": "^4.17.2"
   },
   "peerDependencies": {
     "react": "^0.14.6 || ^15.0.0",
@@ -51,7 +51,7 @@
   },
   "devDependencies": {
     "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.0",
+    "babel-core": "^6.18.2",
     "babel-eslint": "^7.1.0",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-es6-promise": "^1.0.0",
@@ -62,18 +62,18 @@
     "babel-preset-stage-1": "^6.16.0",
     "chai": "^3.5.0",
     "codecov": "^1.0.1",
-    "eslint": "^3.9.0",
+    "eslint": "^3.10.2",
     "eslint-config-standard": "^6.2.1",
     "eslint-config-standard-react": "^4.2.0",
-    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-babel": "^4.0.0",
     "eslint-plugin-promise": "^3.0.0",
-    "eslint-plugin-react": "^6.4.1",
+    "eslint-plugin-react": "^6.7.1",
     "eslint-plugin-standard": "^2.0.1",
-    "isparta": "^4.0.0",
+    "istanbul": "^1.1.0-alpha.1",
     "jsdom": "^9.8.3",
     "mocha": "^3.1.2",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react-addons-test-utils": "^15.4.0",
+    "react-dom": "^15.4.0",
     "rimraf": "^2.5.4"
   },
   "license": "MIT",
