@@ -1,13 +1,16 @@
 import { createStore, compose } from 'redux'
 import rootReducer from './reducer'
 import { firebase as fbConfig } from './config'
-import { reduxReactFirebase } from 'react-redux-firebase'
-// import { reduxFirebase } from 'redux-firebase' // >= v0.1.1
+import { reactReduxFirebase } from 'react-redux-firebase'
 
 export default function configureStore (initialState, history) {
   const createStoreWithMiddleware = compose(
-    reduxReactFirebase(fbConfig, { userProfile: 'users', enableLogging: false }),
-    // reduxFirebase(fbConfig, { userProfile: 'users', enableLogging: false })  // >= v0.1.1
+    reactReduxFirebase(fbConfig,
+      {
+        userProfile: 'users',
+        enableLogging: false
+      }
+    ),
     typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? window.devToolsExtension() : f => f
   )(createStore)
   const store = createStoreWithMiddleware(rootReducer)
