@@ -2,13 +2,18 @@ import { createStore, applyMiddleware, compose } from 'redux'
 import rootReducer from '../reducers'
 import { firebase as fbConfig } from '../config'
 import { syncHistory } from 'react-router-redux'
-import { reduxFirebase } from 'react-redux-firebase' // >= v0.1.1
+import { reactReduxFirebase } from 'react-redux-firebase' // >= v0.1.1
 
 export default function configureStore (initialState, history) {
   const reduxRouterMiddleware = syncHistory(history)
   const createStoreWithMiddleware = compose(
     applyMiddleware(reduxRouterMiddleware),
-    reduxFirebase(fbConfig, { userProfile: 'users', enableLogging: false }),
+    reactReduxFirebase(fbConfig,
+      {
+        userProfile: 'users',
+        enableLogging: false
+      }
+    ),
     typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? window.devToolsExtension() : f => f
   )(createStore)
   const store = createStoreWithMiddleware(rootReducer, initialState)
