@@ -5,7 +5,7 @@ import TodoItem from './TodoItem'
 
 // redux/firebase
 import { connect } from 'react-redux'
-import { firebase, helpers } from 'react-redux-firebase'
+import { firebaseConnect, helpers } from 'react-redux-firebase'
 const { isLoaded, isEmpty, pathToJS, dataToJS } = helpers
 
 class App extends Component {
@@ -15,14 +15,18 @@ class App extends Component {
       push: PropTypes.func.isRequired
     })
   }
-  render () {
-    const { firebase, todos } = this.props
 
-    const handleAdd = () => {
-      const { newTodo } = this.refs
-      firebase.push('/todos', { text: newTodo.value, done: false })
-      newTodo.value = ''
-    }
+  handleAdd = () => {
+    const { firebase } = this.props
+    const { newTodo } = this.refs
+    firebase.push('/todos', { text: newTodo.value, done: false })
+    newTodo.value = ''
+  }
+
+  render () {
+    const { todos } = this.props
+
+    console.log('todos;', todos)
 
     const todosList = (!isLoaded(todos))
                         ? 'Loading'
@@ -50,22 +54,25 @@ class App extends Component {
           {todosList}
           <h4>New Todo</h4>
           <input type='text' ref='newTodo' />
-          <button onClick={handleAdd}>Add</button>
+          <button onClick={this.handleAdd}>
+            Add
+          </button>
         </div>
       </div>
     )
   }
 }
-const fbWrappedComponent = firebase([
-  '/todos'
+const fbWrappedComponent = firebaseConnect([
+  // '/todos'
   // { type: 'once', path: '/todos' } // for loading once instead of binding
   // '/todos#populate=owner:displayNames' // for populating owner parameter from id into string loaded from /displayNames root
-  // '/todos#populate=owner:users' // for populating owner parameter from id to user object loaded from /users root
+  // '/todos#populate=collaborators:users' // for populating owner parameter from id to user object loaded from /users root
+  { path: 'todos', populates: [{ child: 'collaborators', root: 'users' }] } // object notation
   // '/todos#populate=owner:users:displayName' // for populating owner parameter from id within to displayName string from user object within users root
 ])(App)
 
 export default connect(
-  ({firebase}) => ({
+  ({ firebase }) => ({
     todos: dataToJS(firebase, 'todos'),
     profile: pathToJS(firebase, 'profile'),
     auth: pathToJS(firebase, 'auth')
