@@ -1,9 +1,8 @@
-import {
-  SET,
-  NO_VALUE
-} from '../constants'
+import { actionTypes } from '../constants'
+import { promisesForPopulate } from '../utils/populate'
+import { getQueryIdFromPath, applyParamsToQuery } from '../utils/query'
 
-import { map, filter, isString, isObject } from 'lodash'
+const { SET, NO_VALUE } = actionTypes
 
 const getWatchPath = (event, path) =>
   `${event}:${((path.substring(0, 1) === '/') ? '' : '/')}${path}`
@@ -40,29 +39,6 @@ const getWatcherCount = (firebase, event, path, queryId = undefined) => {
 }
 
 /**
- * @description Get query id from query path
- * @param {String} path - Path from which to get query id
- */
-const getQueryIdFromPath = (path) => {
-  const origPath = path
-  let pathSplitted = path.split('#')
-  path = pathSplitted[0]
-
-  let isQuery = pathSplitted.length > 1
-  let queryParams = isQuery ? pathSplitted[1].split('&') : []
-  let queryId = isQuery ? queryParams.map((param) => {
-    let splittedParam = param.split('=')
-    if (splittedParam[0] === 'queryId') {
-      return splittedParam[1]
-    }
-  }).filter(q => q) : undefined
-
-  return (queryId && queryId.length > 0)
-    ? queryId[0]
-    : ((isQuery) ? origPath : undefined)
-}
-
-/**
  * @description Remove/Unset a watcher
  * @param {Object} firebase - Internal firebase object
  * @param {String} event - Type of event to watch for
@@ -91,41 +67,29 @@ const unsetWatcher = (firebase, event, path, queryId = undefined) => {
  * @description Watch a specific event type
  * @param {Object} firebase - Internal firebase object
  * @param {Function} dispatch - Action dispatch function
- * @param {String} event - Type of event to watch for
+ * @param {String} event - Type of event to watch for (defaults to value)
  * @param {String} path - Path to watch with watcher
  * @param {String} dest
  * @param {Boolean} onlyLastEvent - Whether or not to listen to only the last event
  */
-export const watchEvent = (firebase, dispatch, event, path, dest, onlyLastEvent = false) => {
-  let isQuery = false
-  let queryParams = []
-  let queryId = getQueryIdFromPath(path)
-
-  if (queryId) {
-    let pathSplitted = path.split('#')
-    path = pathSplitted[0]
-    isQuery = true
-    queryParams = pathSplitted[1].split('&')
-  }
-
+export const watchEvent = (firebase, dispatch, { type, path, populates, queryParams, queryId, isQuery }, dest, onlyLastEvent = false) => {
   const watchPath = !dest ? path : `${path}@${dest}`
-  const counter = getWatcherCount(firebase, event, watchPath, queryId)
+  const counter = getWatcherCount(firebase, type, watchPath, queryId)
 
   if (counter > 0) {
     if (onlyLastEvent) {
       // listen only to last query on same path
       if (queryId) {
-        unsetWatcher(firebase, event, path, queryId)
+        unsetWatcher(firebase, type, path, queryId)
       } else {
         return
       }
     }
   }
 
-  setWatcher(firebase, event, watchPath, queryId)
+  setWatcher(firebase, type, watchPath, queryId)
 
-  if (event === 'first_child') {
-    // return
+  if (type === 'first_child') {
     return firebase.database()
       .ref()
       .child(path)
@@ -144,56 +108,7 @@ export const watchEvent = (firebase, dispatch, event, path, dest, onlyLastEvent 
   let query = firebase.database().ref().child(path)
 
   if (isQuery) {
-    let doNotParse = false
-
-    queryParams.forEach(param => {
-      param = param.split('=')
-      switch (param[0]) {
-        case 'orderByValue':
-          query = query.orderByValue()
-          doNotParse = true
-          break
-        case 'orderByPriority':
-          query = query.orderByPriority()
-          doNotParse = true
-          break
-        case 'orderByKey':
-          query = query.orderByKey()
-          doNotParse = true
-          break
-        case 'orderByChild':
-          query = query.orderByChild(param[1])
-          break
-        case 'limitToFirst':
-          query = query.limitToFirst(parseInt(param[1], 10))
-          break
-        case 'limitToLast':
-          query = query.limitToLast(parseInt(param[1], 10))
-          break
-        case 'equalTo':
-          let equalToParam = !doNotParse ? parseInt(param[1], 10) || param[1] : param[1]
-          equalToParam = equalToParam === 'null' ? null : equalToParam
-          query = param.length === 3
-            ? query.equalTo(equalToParam, param[2])
-            : query.equalTo(equalToParam)
-          break
-        case 'startAt':
-          let startAtParam = !doNotParse ? parseInt(param[1], 10) || param[1] : param[1]
-          startAtParam = startAtParam === 'null' ? null : startAtParam
-          query = param.length === 3
-            ? query.startAt(startAtParam, param[2])
-            : query.startAt(startAtParam)
-          break
-        case 'endAt':
-          let endAtParam = !doNotParse ? parseInt(param[1], 10) || param[1] : param[1]
-          endAtParam = endAtParam === 'null' ? null : endAtParam
-          query = param.length === 3
-            ? query.endAt(endAtParam, param[2])
-            : query.endAt(endAtParam)
-          break
-        default:
-          break
-      } })
+    query = applyParamsToQuery(queryParams, query)
   }
 
   const runQuery = (q, e, p, params) => {
@@ -201,11 +116,15 @@ export const watchEvent = (firebase, dispatch, event, path, dest, onlyLastEvent 
     if (e === 'once') {
       return q.once('value')
         .then(snapshot =>
-          dispatch({ type: SET, path, data: snapshot.val() })
+          dispatch({
+            type: SET,
+            path,
+            data: snapshot.val()
+          })
         )
     }
-
     // Handle all other queries
+    // TODO: Handle onError of query (i.e. security errors)
     q.on(e, snapshot => {
       let data = (e === 'child_removed') ? undefined : snapshot.val()
       const resultPath = dest || (e === 'value') ? p : `${p}/${snapshot.key}`
@@ -217,13 +136,8 @@ export const watchEvent = (firebase, dispatch, event, path, dest, onlyLastEvent 
         }
       }
 
-      // Get list of populates
-      const populates = filter(params, param =>
-        param.indexOf('populate') !== -1
-      ).map(p => p.split('=')[1])
-
-      // Dispatch standard if no populates
-      if (!populates || !populates.length) {
+      // Dispatch standard event if no populates exists
+      if (!populates) {
         return dispatch({
           type: SET,
           path: resultPath,
@@ -232,59 +146,9 @@ export const watchEvent = (firebase, dispatch, event, path, dest, onlyLastEvent 
         })
       }
 
-      // TODO: Handle object based populates
-      // TODO: Handle multiple populates
-      // Handle First Populate
-      const populate = populates[0]
-      const listToPopulate = snapshot.val()
-      const paramToPopulate = populate.split(':')[0]
-      const populateRoot = populate.split(':')[1]
-      const populateKey = populate.split(':')[2]
-      const listRef = firebase.database().ref().child(populateRoot)
+      // TODO: Allow setting of unpopulated data before starting population through config
 
-      // Create list of promises (one for each population)
-      const promises = map(listToPopulate, (item, key) => {
-        if (!item[paramToPopulate]) {
-          return Object.assign(item, { _key: key })
-        }
-
-        // TODO: Handle populating a list
-        return !isString(item[paramToPopulate])
-            // Parameter to be populated is not an id
-            ? Promise.reject(`
-                Population id is not a string.\n
-                Type: ${typeof item[paramToPopulate]}\n
-                Id: ${JSON.stringify(item[paramToPopulate])}
-              `)
-            : listRef.child(item[paramToPopulate])
-                .once('value')
-                .then(snap =>
-                  // Handle population value not existing
-                  !snap.val()
-                    ? item[paramToPopulate]
-                    // Handle population value (object/string)
-                    : isObject(snap.val())
-                      // Handle selecting of a specific value within object
-                      ? (populateKey && snap.val()[populateKey])
-                        // Return value at populate key
-                        ? snap.val()[populateKey]
-                        // Return object with snap and key attached
-                        : Object.assign(
-                            snap.val(),
-                            { _snap: snap, _key: snap.key }
-                          )
-                      // Return value (string, number or bool)
-                      : snap.val()
-
-                )
-                .then((populatedList) => {
-                  const newItem = item
-                  newItem[paramToPopulate] = populatedList
-                  return Object.assign(newItem, { _key: key })
-                })
-      })
-
-      Promise.all(promises)
+      promisesForPopulate(firebase, data, populates)
         .then((list) => {
           dispatch({
             type: SET,
@@ -295,7 +159,7 @@ export const watchEvent = (firebase, dispatch, event, path, dest, onlyLastEvent 
     })
   }
 
-  runQuery(query, event, path, queryParams)
+  runQuery(query, type, path, queryParams)
 }
 
 /**
@@ -315,7 +179,7 @@ export const unWatchEvent = (firebase, event, path, queryId = undefined) =>
  */
 export const watchEvents = (firebase, dispatch, events) =>
     events.forEach(event =>
-      watchEvent(firebase, dispatch, event.name, event.path)
+      watchEvent(firebase, dispatch, event)
     )
 
 /**
@@ -325,7 +189,7 @@ export const watchEvents = (firebase, dispatch, events) =>
  */
 export const unWatchEvents = (firebase, events) =>
     events.forEach(event =>
-      unWatchEvent(firebase, event.name, event.path)
+      unWatchEvent(firebase, event.type, event.path)
     )
 
 export default { watchEvents, unWatchEvents }
